+++ src/components/beta/gux-chart-line/gux-chart-line.tsx
@@ -31,7 +31,8 @@ export class GuxLineChart {
     },
     config: {
       axis: {
-        ticks: false
+        ticks: false,
+        titlePadding: 8
       },
       axisX: {
         labelAngle: 0
@@ -69,6 +70,18 @@ export class GuxLineChart {
 
   @Prop()
   includeLegend: boolean;
+
+  @Prop()
+  legendPosition:
+    | 'left'
+    | 'right'
+    | 'top'
+    | 'bottom'
+    | 'top-left'
+    | 'top-right'
+    | 'bottom-left'
+    | 'bottom-right'
+    | 'none' = 'right';
 
   @Prop()
   includeDataPointMarkers: boolean;
@@ -137,6 +150,10 @@ export class GuxLineChart {
 
     if (this.includeLegend) {
       this.baseChartSpec.encoding.color.legend = true;
+    }
+
+    if (this.legendPosition) {
+      this.baseChartSpec.config.legend.orient = this.legendPosition;
     }
 
     const xFieldName = this.xFieldName;